Moving teletry folder to /lib
