+++ src/api/error/ErrorReport.ts
@@ -10,7 +10,7 @@ import * as pkg from '../../../package.json'
 import logger from '../logger'
 import { SessionManager } from '../session/SessionManager'
 import { getPlatform } from '../../lib/utils/getPlatform'
-import { TelemetryCollector } from '../telemetry/TelemetryCollector'
+import { TelemetryCollector } from '../../lib/telemetry/TelemetryCollector'
 import { CommandError } from './errors'
 
 interface CustomErrorReportCreateArgs extends ErrorReportCreateArgs {
