+++ src/api/index.ts
@@ -3,7 +3,6 @@ export * from './error'
 export * from './metrics'
 export * from './oclif'
 export * from './session'
-export * from './telemetry'
 export * from './manifest'
 export * from './env'
 export * from './files'
