+++ src/api/oclif/CustomCommand.ts
@@ -2,7 +2,7 @@ import { hrTimeToMs } from '../../lib/utils/hrTimeToMs'
 import { Metric } from '../metrics/MetricReport'
 import { onError } from '../../oclif/hooks/init'
 import { ParsingToken } from '@oclif/parser/lib/parse'
-import { TelemetryCollector } from '../telemetry/TelemetryCollector'
+import { TelemetryCollector } from '../../lib/telemetry/TelemetryCollector'
 import { TraceConfig } from '../../lib/globalConfigs/traceConfig'
 import * as Parser from '@oclif/parser'
 import OclifCommand, { flags as oclifFlags } from '@oclif/command'
