+++ src/lib/telemetry/TelemetryCollector.ts
@@ -3,10 +3,10 @@ import { randomBytes } from 'crypto'
 import { ensureFileSync, writeJsonSync } from 'fs-extra'
 import { join } from 'path'
 import * as pkgJson from '../../../package.json'
-import logger from '../logger'
+import logger from '../../api/logger'
 import { PathConstants } from '../../lib/constants/Paths'
-import { ErrorReport } from '../error/ErrorReport'
-import { Metric, MetricReport, MetricReportObj } from '../metrics/MetricReport'
+import { ErrorReport } from '../../api/error/ErrorReport'
+import { Metric, MetricReport, MetricReportObj } from '../../api/metrics/MetricReport'
 import { spawnUnblockingChildProcess } from '../../lib/utils/spawnUnblockingChildProcess'
 import { ITelemetryLocalStore, TelemetryLocalStore } from './TelemetryStore'
