+++ src/lib/telemetry/TelemetryReporter/PendingTelemetryDataManager.ts
@@ -2,8 +2,8 @@ import { randomBytes } from 'crypto'
 import { ensureDir, move, readdir, remove, stat, writeJson } from 'fs-extra'
 import glob from 'globby'
 import { join } from 'path'
-import { TelemetryMetaMetrics } from '../../metrics/MetricNames'
-import { MetricReport } from '../../metrics/MetricReport'
+import { TelemetryMetaMetrics } from '../../../api/metrics/MetricNames'
+import { MetricReport } from '../../../api/metrics/MetricReport'
 import { FileLock } from '../../../lib/utils/FileLock'
 import { hrTimeToMs } from '../../../lib/utils/hrTimeToMs'
 import { TelemetryCollector, TelemetryFile } from '../TelemetryCollector'
